refactor employee model
refactor employee model
